changed default type to double

+++ src/constants/type.h
@@ -2,7 +2,7 @@
 #define CONSTANTS_TYPE_H
 
 /// floating point precision (e.g., double or float)
-using fpp = float;
+using fpp = double;
 
 // ensure fpp is either double or float to use correct HDF5 and MPI type aliases
 static_assert(std::is_same_v<fpp, double> || std::is_same_v<fpp, float>);
